修改AcquirePort相关文件 详见triclops的pr #27
